+++ src/device/device/VI/Settings/SystemSubsystemSettings.cs
@@ -66,7 +66,7 @@ public class SystemSubsystemSettings() : System.ComponentModel.INotifyPropertyCh
     /// <returns>   <see langword="true"/> if it succeeds; otherwise, <see langword="false"/>. </returns>
     protected bool SetProperty<T>( T oldValue, T newValue, Action callback, [System.Runtime.CompilerServices.CallerMemberName] string? propertyName = null )
     {
-#if NET8_0_OR_GREATER
+#if NET5_0_OR_GREATER
         ArgumentNullException.ThrowIfNull( callback, nameof( callback ) );
 #else
         if ( callback is null ) throw new ArgumentNullException( nameof( callback ) );