+++ src/vi/k2600/k2600.mstest/Measure/ResistanceSettings.cs
@@ -62,7 +62,7 @@ public class ResistanceSettings() : System.ComponentModel.INotifyPropertyChanged
     /// <returns>   <see langword="true"/> if it succeeds; otherwise, <see langword="false"/>. </returns>
     protected bool SetProperty<T>( T oldValue, T newValue, Action callback, [System.Runtime.CompilerServices.CallerMemberName] string? propertyName = null )
     {
-#if NET8_0_OR_GREATER
+#if NET5_0_OR_GREATER
         ArgumentNullException.ThrowIfNull( callback, nameof( callback ) );
 #else
         if ( callback is null ) throw new ArgumentNullException( nameof( callback ) );