+++ changelog.md
@@ -2,9 +2,13 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9447]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9450]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9450] - 2025-11-15
+- Transition .NET 9.0 projects to .NET 10 and VS2026.
+- Move the list of ISR Visa resources from the protected public documents folder to the application data.
 
 ## [8.1.9447] - 2025-11-12
 Script Loader